. Make rlen a size_t rather than a int.  This is the correct argument type   to pass to sysctl and fixes a crash on 64 bit architectures where size_t   is 64 bits rather than 32 bits.

+++ src/solaris/native/com/sun/management/UnixOperatingSystem_md.c
@@ -292,7 +292,7 @@ Java_com_sun_management_UnixOperatingSystem_getTotalPhysicalMemorySize
 #ifdef _ALLBSD_SOURCE
     jlong result;
     int mib[2];
-    int rlen;
+    size_t rlen;
 
     mib[0] = CTL_HW;
     mib[1] = HW_PHYSMEM;
